+++ flaskblog/routes.py
@@ -1,10 +1,8 @@
-from flask import Flask, render_template, url_for, flash, redirect
-from forms import RegistrationForm, LoginForm
+from flask import  render_template, url_for, flash, redirect
+from flaskblog.forms import RegistrationForm, LoginForm
+from flaskblog.models import User, Post
+from flaskblog import app
 
-app = Flask(__name__)
-app.config['SECRET_KEY'] = '5791628bb0b13ce0c676dfde280ba245'
-#got using secrets module
-#dummy data
 posts = [
     {
         'author': 'Corey Schafer',
@@ -19,7 +17,6 @@ posts = [
         'date_posted': 'April 21, 2018'
     }
 ]
-
 
 @app.route("/")
 @app.route("/home")
@@ -51,7 +48,3 @@ def login():
         else:
             flash('Login Unsuccessful. Please check username and password', 'danger')
     return render_template('login.html', title='Login', form=form)
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
