+++ requirements.txt
@@ -1,8 +1,16 @@
 click==7.1.2
+DateTime==4.3
+dnspython==1.16.0
+email-validator==1.1.1
 Flask==1.1.2
 Flask-SQLAlchemy==2.4.1
+Flask-WTF==0.14.3
+idna==2.9
 itsdangerous==1.1.0
 Jinja2==2.11.2
 MarkupSafe==1.1.1
+pytz==2020.1
 SQLAlchemy==1.3.17
 Werkzeug==1.0.1
+WTForms==2.3.1
+zope.interface==5.1.0
