+++ run.py
@@ -1,0 +1,8 @@
+from flaskblog import app
+
+
+
+
+
+if __name__ == '__main__':
+    app.run(debug=True)
